JC first stab at cookie stuff
