+++ src/widgets/Widget.ts
@@ -5,22 +5,7 @@ import AnalyticsApi, { SQHDetails } from "../api/AnalyticsApi";
 import WidgetApi from "../api/WidgetApi";
 import { WidgetType, WidgetContext } from "../types";
 import { isObject, hasProps } from "../utils/validate";
-
-function readCookie(cname) {
-  var name = cname + "=";
-  var decodedCookie = decodeURIComponent(document.cookie);
-  var ca = decodedCookie.split(';');
-  for(var i = 0; i <ca.length; i++) {
-    var c = ca[i];
-    while (c.charAt(0) == ' ') {
-      c = c.substring(1);
-    }
-    if (c.indexOf(name) == 0) {
-      return c.substring(name.length, c.length);
-    }
-  }
-  return "";
-}
+import readCookie from "../utils/readCookie";
 
 /** @hidden */
 const _log = debug("squatch-js:widget");
@@ -46,7 +31,6 @@ export default abstract class Widget {
   frame: HTMLIFrameElement;
   type: WidgetType;
   content: string;
-  squatchCookie: string;
   analyticsApi: AnalyticsApi;
   widgetApi: WidgetApi;
   context: WidgetContext;
@@ -57,7 +41,6 @@ export default abstract class Widget {
     this.content =
       params.content === "error" ? this._error(params.rsCode) : params.content;
     this.type = params.type;
-    this.squatchCookie = readCookie("refParam");
     this.widgetApi = params.api;
     this.npmCdn = params.npmCdn;
     this.analyticsApi = new AnalyticsApi({ domain: params.domain });
@@ -248,14 +231,14 @@ export default abstract class Widget {
         lastName: lastName || null,
         id: this.context.user.id,
         accountId: this.context.user.accountId,
-        cookie: this.squatchCookie
+        cookie: readCookie("saasquatchCookie")
       };
 
       response = this.widgetApi.upsertUser({
         user: userObj,
         engagementMedium,
         widgetType: this.type,
-        cookie: this.squatchCookie,
+        cookie: readCookie("saasquatchCookie"),
         jwt,
       });
     } else if (this.context.type === "cookie") {
